requirements.txt
sounddevice
google-generativeai
scipy
elevenlabs